% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{as.list,Competition-method}
\alias{as.list,Competition-method}
\title{Transform a Competition into a list}
\usage{
\S4method{as.list}{Competition}(x)
}
\arguments{
\item{x}{An object of class Competition}
}
\description{
Transform a Competition into a list
}
