% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{event_score,Event-method}
\alias{event_score,Event-method}
\title{Calculate the score result of an event}
\usage{
\S4method{event_score}{Event}(x)
}
\arguments{
\item{x}{An object to calculate the score from}
}
\description{
Calculate the score result of an event
}
