% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routines.R
\name{get_routines,Event-method}
\alias{get_routines,Event-method}
\title{Get the routines of a team in an event}
\usage{
\S4method{get_routines}{Event}(x, team)
}
\arguments{
\item{x}{An object to calculate the score from}

\item{team}{Which team the routines should be extracted for}
}
\description{
Get the routines of a team in an event
}
