% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routines.R
\name{get_routines,Pairing-method}
\alias{get_routines,Pairing-method}
\title{Get the routine of a team in a pairing}
\usage{
\S4method{get_routines}{Pairing}(x, team)
}
\arguments{
\item{x}{An object to calculate the score from}

\item{team}{Which team the routines should be extracted for}
}
\description{
Get the routine of a team in a pairing
}
