% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gymnast.R
\docType{class}
\name{Gymnast-class}
\alias{Gymnast-class}
\title{An S4 class that stores a Gymnast}
\description{
An S4 class that stores a Gymnast
}
\section{Slots}{

\describe{
\item{\code{id}}{Numeric ID as on homepage}

\item{\code{name}}{Name of the gymnast}

\item{\code{nationality}}{Nationality of the gymnast}
}}

