% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_starts.R
\name{home_starts,Pairing-method}
\alias{home_starts,Pairing-method}
\title{Calculate the score result of a pairing}
\usage{
\S4method{home_starts}{Pairing}(x)
}
\arguments{
\item{x}{An object to calculate the score from}
}
\description{
Calculate the score result of a pairing
}
