% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_starts.R
\name{home_starts<-}
\alias{home_starts<-}
\title{Set the starting team of the object}
\usage{
home_starts(x) <- value
}
\arguments{
\item{x}{The object to set}

\item{value}{The value that should be set to (boolean)}
}
\description{
Set the starting team of the object
}
