% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{max,Events-method}
\alias{max,Events-method}
\title{Get the Event with the hightest Score difference for the home team}
\usage{
\S4method{max}{Events}(x)
}
\arguments{
\item{x}{Events object to choose from}
}
\description{
Get the Event with the hightest Score difference for the home team
}
