% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{min,Events-method}
\alias{min,Events-method}
\title{Get the Event with the lowest Score difference for the home team}
\usage{
\S4method{min}{Events}(x)
}
\arguments{
\item{x}{Events object to choose from}
}
\description{
Get the Event with the lowest Score difference for the home team
}
