\docType{class}
\name{Pairing-class}
\alias{Pairing-class}
\title{An S4 class that stores a Pairing}
\description{
An S4 class that stores a Pairing
}
\section{Slots}{

\describe{
\item{\code{home}}{Routine of the home team}

\item{\code{guest}}{Routine of the guest team}

\item{\code{home_starts}}{logical which team started. TRUE for home}

}}
