% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{rating,Event-method}
\alias{rating,Event-method}
\title{Rate the score result of an event}
\usage{
\S4method{rating}{Event}(x)
}
\arguments{
\item{x}{An object to rate the score}
}
\description{
Rate the score result of an event
}
