% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{rating,Pairing-method}
\alias{rating,Pairing-method}
\title{Calculate the score result of a pairing}
\usage{
\S4method{rating}{Pairing}(x)
}
\arguments{
\item{x}{An object to calculate the score from}
}
\description{
Calculate the score result of a pairing
}
