% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_score.R
\name{rating}
\alias{rating}
\title{Rate the score results}
\usage{
rating(x, ...)
}
\arguments{
\item{x}{An object to calculate the rating from}

\item{...}{Extra named arguments}
}
\description{
Rate the score results
}
