\docType{class}
\name{Routine-class}
\alias{Routine-class}
\title{An S4 class that stores a Routine}
\description{
An S4 class that stores a Routine
}
\section{Slots}{

\describe{
\item{\code{gymnast}}{gymnast responsible for routine}

\item{\code{dvalue}}{dvalue of the routine}

\item{\code{endvalue}}{endvalue of the routine}

}}
