% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score,Competition-method}
\alias{score,Competition-method}
\title{Calculate the score result of a competition}
\usage{
\S4method{score}{Competition}(x)
}
\arguments{
\item{x}{An object to calculate the score from}
}
\description{
Calculate the score result of a competition
}
