% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score,Pairing-method}
\alias{score,Pairing-method}
\title{Calculate the score result of a pairing}
\usage{
\S4method{score}{Pairing}(x)
}
\arguments{
\item{x}{An object to calculate the score from}
}
\description{
Calculate the score result of a pairing
}
