% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_score.R
\name{score}
\alias{score}
\title{Calculate the score results}
\usage{
score(x, ...)
}
\arguments{
\item{x}{An object to calculate the score from}

\item{...}{Extra named arguments}
}
\description{
Calculate the score results
}
