% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_diff,Event-method}
\alias{score_diff,Event-method}
\title{Rate the score result of an event}
\usage{
\S4method{score_diff}{Event}(x)
}
\arguments{
\item{x}{An object to rate the score}
}
\description{
Rate the score result of an event
}
