% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_score.R
\name{score_diff}
\alias{score_diff}
\title{Rate the score results}
\usage{
score_diff(x, ...)
}
\arguments{
\item{x}{An object to calculate the score_diff from}

\item{...}{Extra named arguments}
}
\description{
Rate the score results
}
