\name{scorehelper-package}
\alias{scorehelper-package}
\alias{scorehelper}
\docType{package}
\title{\packageTitle{scorehelper}}
\description{\packageDescription{scorehelper}}
\section{Package Content}{\packageIndices{scorehelper}}
\author{\packageAuthor{scorehelper}}
\section{Maintainer}{\packageMaintainer{scorehelper}}
\keyword{package}
