% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_starts.R
\name{home_starts}
\alias{home_starts}
\title{Set the starting team in an object}
\usage{
home_starts(x)
}
\arguments{
\item{x}{An object to set the starting team to}

\item{home_starts}{Boolean if home starts'}
}
\description{
Set the starting team in an object
}
