% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gymnast.R
\docType{class}
\name{Team-class}
\alias{Team-class}
\title{An S4 class that stores Information about a Team}
\description{
An S4 class that stores Information about a Team
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the team}
}}

