% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorehelper-package.R
\docType{package}
\name{scorehelper-package}
\alias{scorehelper}
\alias{scorehelper-package}
\title{scorehelper: What the Package Does Using Title Case}
\description{
More details about what the package does. See \url{http://cran.r-project.org/doc/manuals/r-release/R-exts.html#The-DESCRIPTION-file} for details on how to write this part.
}
\keyword{internal}
